tests: refactor internals tests to not depend on each other
