+++ src/python/pants/engine/internals/synthetic_targets_test.py
@@ -17,7 +17,6 @@ from pants.base.specs import RawSpecsWithoutFileOwners, RecursiveGlobSpec
 from pants.core.target_types import FileTarget, GenericTarget, LockfilesGeneratorTarget
 from pants.engine.addresses import Address, Addresses
 from pants.engine.environment import EnvironmentName
-from pants.engine.internals.specs_rules_test import resolve_raw_specs_without_file_owners
 from pants.engine.internals.synthetic_targets import (
     SyntheticAddressMaps,
     SyntheticTargetsRequest,
@@ -25,6 +24,7 @@ from pants.engine.internals.synthetic_targets import (
     rules,
 )
 from pants.engine.internals.target_adaptor import TargetAdaptor as _TargetAdaptor
+from pants.engine.internals.testutil import resolve_raw_specs_without_file_owners
 from pants.engine.rules import QueryRule, rule
 from pants.engine.target import (
     Dependencies,
